refactor: used custom styles file for the farm pages and widgets
