--- lib/pages/farm/modules/box_decoration.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-BoxDecoration boxDecoration(
-  Color col,
-  double rad,
-  double borWid,
-  Color borCol,
-) {
-  return BoxDecoration(
-    color: col,
-    borderRadius: BorderRadius.circular(rad),
-    border: Border.all(color: borCol, width: borWid),
-  );
-}
