+++ lib/pages/farm/widgets/double_logo_farm_title.dart
@@ -2,9 +2,8 @@
 
 import 'package:ax_dapp/pages/farm/dialogs/dual_stake_dialog.dart';
 import 'package:ax_dapp/pages/farm/dialogs/stake_dialog.dart';
-import 'package:ax_dapp/pages/farm/modules/box_decoration.dart';
-import 'package:ax_dapp/pages/farm/modules/page_text_style.dart';
 import 'package:ax_dapp/service/controller/farms/farm_controller.dart';
+import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:flutter/material.dart';
 
 Widget doubleLogoFarmTitle(
@@ -44,7 +43,7 @@ Widget doubleLogoFarmTitle(
         Expanded(
           child: Text(
             '${farm.athlete!} Farm',
-            style: textStyle(Colors.white, 16, false, false),
+            style: textStyle(Colors.white, 16, isBold: false),
           ),
         ),
         Container(
@@ -80,7 +79,7 @@ Widget doubleLogoFarmTitle(
             ),
             child: Text(
               'Stake',
-              style: textStyle(Colors.black, 14, true, false),
+              style: textStyle(Colors.black, 14, isBold: true),
             ),
           ),
         ),
