+++ lib/pages/farm/widgets/farm_item.dart
@@ -1,9 +1,8 @@
 // ignore_for_file: avoid_positional_boolean_parameters
 
-import 'package:ax_dapp/pages/farm/components/single_logo_farm_title.dart';
-import 'package:ax_dapp/pages/farm/modules/box_decoration.dart';
-import 'package:ax_dapp/pages/farm/modules/page_text_style.dart';
+import 'package:ax_dapp/pages/farm/widgets/single_logo_farm_title.dart';
 import 'package:ax_dapp/service/controller/farms/farm_controller.dart';
+import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
@@ -21,7 +20,7 @@ Widget farmItem(
   if (cardHeight < minCardHeight) cardHeight = minCardHeight;
   if (cardHeight > maxCardHeight) cardHeight = maxCardHeight;
 
-  final txStyle = textStyle(Colors.grey[600]!, 14, false, false);
+  final txStyle = textStyle(Colors.grey[600]!, 14, isBold: false);
   Widget farmTitleWidget;
   farmTitleWidget = singleLogoFarmTitle(context, isWeb, farm, cardWidth);
