+++ lib/pages/farm/widgets/single_logo_farm_title.dart
@@ -1,13 +1,12 @@
 // ignore_for_file: avoid_positional_boolean_parameters
 
-import 'package:ax_dapp/pages/farm/components/sport_token.dart';
 import 'package:ax_dapp/pages/farm/dialogs/stake_dialog.dart';
-import 'package:ax_dapp/pages/farm/modules/box_decoration.dart';
-import 'package:ax_dapp/pages/farm/modules/page_text_style.dart';
+import 'package:ax_dapp/pages/farm/widgets/sport_token.dart';
 import 'package:ax_dapp/service/controller/farms/farm_controller.dart';
+import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:ax_dapp/util/toast_extensions.dart';
-import 'package:flutter/material.dart';
 import 'package:ax_dapp/wallet/bloc/wallet_bloc.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 Widget singleLogoFarmTitle(
@@ -70,7 +69,7 @@ Widget singleLogoFarmTitle(
             },
             child: Text(
               'Stake',
-              style: textStyle(Colors.black, 14, true, false),
+              style: textStyle(Colors.black, 14, isBold: true),
             ),
           ),
         ),
