+++ lib/pages/farm/widgets/stake_approve_button.dart
